Fixed login button (#51), resume now doesn't auto-scroll to page (#48), added tooltip for more info (#49).
Part of the issue with the login button was that authentication wasn't hapenning first which means we weren't getting session data when processing EJS files through Harp. Now session / user data is being correctly passed to EJS so that we can show user information. Currently as a placeholder it shows the Google profile image.